Moved to shims, to prevent breakage with Spark 3.4.
