+++ sql-plugin/src/main/311until340-all/scala/org/apache/spark/sql/hive/rapids/shims/HiveProviderCmdShims.scala
@@ -21,7 +21,6 @@ import com.nvidia.spark.rapids.shims.OptimizedCreateHiveTableAsSelectCommandMeta
 
 import org.apache.spark.sql.execution.command.{DataWritingCommand, RunnableCommand}
 import org.apache.spark.sql.hive.execution.{InsertIntoHiveTable, OptimizedCreateHiveTableAsSelectCommand}
-import org.apache.spark.sql.hive.rapids.GpuInsertIntoHiveTableMeta
 
 trait HiveProviderCmdShims extends HiveProvider {
